+++ RELEASE_NOTES.md
@@ -1,2 +1,6 @@
 ### 0.0.1 - July 17 2017
 * Initial release
+
+### 0.02 - July 22 2017
+* Added a small C helper to retrieve architecture details
+* Added support for x86 instruction details